+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### Octonica.ClickHouseClient Next Version, Unscheduled
+
+#### New Feature
+
+* Add method `ClickHouseDataReader.ConfigureColumnReader` which sets a value cast callback function for the column.
+ The callback function is invoked whenever the reader reads a non-null value from the column. This feature could be useful in
+ cases when ClcikHouseClient doesn't have a built-in type conversion rule, for expample, `double -> decimal`, `Guid -> string` or `uint -> int`.
+
 ### Octonica.ClickHouseClient v2.2.11, 2023-01-11
 
 #### Bug Fix